projects/01-doc-qa-system/requirements.txt
# 智能文档问答系统依赖

# LangChain 核心
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1

# 向量数据库
faiss-cpu>=1.7.4

# 文档解析
pypdf>=3.17.0
python-docx>=1.1.0
markdown>=3.5.0

# Web UI
gradio>=4.0.0

# 工具
python-dotenv>=1.0.0
numpy>=1.24.0
